requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.0
Django==5.1.5
djangorestframework==3.15.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
